Added Access-Control-Allow-Origin: * header
I'm working on a thing that'll let you preview CSS customizations to your scrapbook in real-time by injecting the CSS into an iframe, but it didn't work cause of CORS restrictions. I have a workaround that's kinda working for now, but it'd be nice to have this header set instead.