Proper tooltip display with react-tooltip
The `titleOnValue` property in `react-calendar-heatmap` only works with
the third party bootstrap tooltip library since it sets a `title`
attribute instead of the SVG standard `<title>` tag. In any case, it
appears the preferred method to implement the tooltip is with
`react-tooltip` and the `tooltipDataAttrs` method instead as seen
in https://github.com/kevinsqi/react-calendar-heatmap/blob/master/demo/src/Demo.js

`react-tooltip` doesn't play nicely with server side rendering,
so we use a next/dynamic import to disable SSR for the `ReactTooltip`
component.